+++ Assets/VRM10_Samples/VRM10Viewer/VRM10Motion.cs
@@ -1,0 +1,7 @@
+namespace UniVRM10.VRM10Viewer
+{
+    public class VRM10Motion
+    {
+        
+    }
+}
