+++ Assets/VRM10_Samples/VRM10Viewer/VRM10Motion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 464a0c26baba19442ac1892219f2bb2e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
